tests/wpt/web-platform-tests/bluetooth/characteristic/service-same-object.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = '[SameObject] test for BluetoothRemoteGATTCharacteristic ' +
  'service.';

bluetooth_test(() => getMeasurementIntervalCharacteristic()
    .then(({characteristic}) =>
        assert_equals(characteristic.service, characteristic.service)),
    test_desc);
</script>
